+++ hyperwallet/tests/test_models.py
@@ -11,6 +11,8 @@ from hyperwallet import (
     BankCard,
     PrepaidCard,
     PaperCheck,
+    Transfer,
+    PayPalAccount,
     Payment,
     Balance,
     Receipt,
@@ -216,6 +218,42 @@ class ModelTest(unittest.TestCase):
         self.assertEqual(
             test_paper_check.__repr__(),
             'PaperCheck({date}, {token})'.format(
+                date=self.transfer_method_data.get('createdOn'),
+                token=self.transfer_method_data.get('token')
+            )
+        )
+
+    '''
+
+    Transfer
+
+    '''
+
+    def test_transfer_model(self):
+
+        test_transfer = Transfer(self.transfer_method_data)
+
+        self.assertEqual(
+            test_transfer.__repr__(),
+            'Transfer({date}, {token})'.format(
+                date=self.transfer_method_data.get('createdOn'),
+                token=self.transfer_method_data.get('token')
+            )
+        )
+
+    '''
+
+    PayPal Account
+
+    '''
+
+    def test_paypal_account_model(self):
+
+        test_paypal_account = PayPalAccount(self.transfer_method_data)
+
+        self.assertEqual(
+            test_paypal_account.__repr__(),
+            'PayPalAccount({date}, {token})'.format(
                 date=self.transfer_method_data.get('createdOn'),
                 token=self.transfer_method_data.get('token')
             )