+++ CHANGELOG.rst
@@ -1,12 +1,17 @@
 Changelog
 =========
 
-1.1.3 (current)
+1.1.4 (current)
 ------------------
 
-- Added Layer 7 encryption for Hyperwallet client
+- Added PayPal account endpoint
 
-1.1.2 (2018-08-03)
+1.1.3 (2018-07-05)
+------------------
+
+- Added transfer endpoint
+
+1.1.2 (2018-03-20)
 ------------------
 
 - Added bank card endpoint